lol uh my cop will be gone soon ;-;

+++ src/index.ts
@@ -1,6 +1,10 @@
 /* eslint-disable linebreak-style */
 // Requires as imports up-top
-const path = require('path');
+import { Bot } from 'aoi.js';
+import {Request, Response} from 'express';
+import path from 'path';
+import { DashOptions, DashInterface, ThemeFlags } from "../index";
+// const path = require('path');
 const fs = require('fs');
 const csrf = require("csurf");
 const Api = require('./api');
@@ -13,19 +17,21 @@ const Theme = require('./themes');
 const debug = require('debug')('aoijs.panel:main');
 const {DiscordClient}= require('./discord_handle.js');
 require('./version_check');
-class Dash {
+
+class Dash implements DashInterface {
   static Session = sessions;
-  constructor(ops) {
+    constructor(ops: DashOptions) {
     let name = ops;
     let bot = name.bot;
     let port = name.port;
     let username = name.username;
-    let pass = name.password;
+    let pass = name.pass;
     let command = name.command;
     this.ops = name;
 this.ownersIds = ops.owners || [];
 if(this.ownersIds.length < 0) console.warn("No owners specified, nobody can access commands!")
 this.information  = ops.information.homepage;
+//@ts-ignore
 bot.panel = this;
     this.bot = bot
 this.port = port
@@ -38,9 +44,21 @@ if(ops.express?.store) this.store = ops.express?.store;
       this.theme = ops.theme || Theme.FLAGS.default;
 this.discord = new DiscordClient(ops.discord, this)
   }
-  renderFile(path, parms, cb) {
+  store: any;
+  cmd: string;
+  api: Api;
+  theme: ThemeFlags;
+  discord: DiscordInterface;
+  information: { body: { title: string; description: string; }[]; };
+  port: number;
+  username: string;
+  pass: string;
+  ownersIds: string[];
+  ops: DashOptions;
+  bot: Bot;
+  private renderFile(path:string, parms?:any, cb?: Function) {
  return new Promise((resolve, rej) => {
-  const done = (err, res) => {
+  const done = (err:string | Error| null, res:any) => {
     if(err && cb) {
       cb(err, null)
     // Promise.reject(err)
@@ -62,7 +80,7 @@ rej(err)
 done(null, res)
  })
   }
-start() {
+public start() {
 const bot = this.bot
 const port = this.port
 const command = this.cmd
@@ -87,10 +105,12 @@ const SESSION_SECRET = encodeURI("7g2rf382vf8y2vcy8vc8yev8cyv28yvcy8evcw8yvc1&%$
 app.use(csrf({ cookie: true }));
     app.use(this.discord.session())
     app.use('/api/', this.api.app)
-    app.get('/auth/discord/redirect', (req,res) => res.redirect(/* 200, */ this.discord.GetAuthUrl()))
-    app.get('/command/edit', islogin, async function(req,res) {
-        let pathh = req.query.path
-        let name = pathh.replace(/%2F/g, '/')
+    app.get('/auth/discord/redirect', (req: Request, res: Response) => res.redirect(/* 200, */ this.discord.GetAuthUrl()))
+    app.get('/command/edit', islogin, async function(req: Request, res: Response) {
+       
+      let pathh:string | null = req.query.path as string;
+      if(!pathh) return res.status(500).end();
+      let name = pathh.replace(/%2F/g, '/')
         pathh = pathh.replace(/%2F/g, ',')
         let code = fs.readFileSync(path.join(process.cwd(), pathh))
         
@@ -99,7 +119,7 @@ app.use(csrf({ cookie: true }));
     
     
     
-    app.post('/command/save', islogin, async function(req,res) {
+    app.post('/command/save', islogin, async function(req: Request, res: Response) {
         let name = req.body.path
        name = name.replace(/\//g, path.sep)
         let nowname = command + path.sep + req.body.name
@@ -112,23 +132,25 @@ app.use(csrf({ cookie: true }));
     })
     
     
-    app.get('/login', login,async function(req,res) {
+    app.get('/login', login,async function(req: Request, res: Response) {
 
         res.send(await _this.renderFile(path.join(__dirname, "views", "login.ejs")))
         
         })
-        app.get('/', async function(req,res) {
+        app.get('/', async function(req: Request, res: Response) {
 res.send(await _this.renderFile(path.join(Theme.getPath(_this.theme), "index.ejs"),{ info: _this.information, bot }) )
         })
     
-    app.get('/auth/discord', async (req,res,next) => await this.discord.authorize(req,res,next), (req,res) => {
+    app.get('/auth/discord', async (req: Request, res: Response, next: Function) => await this.discord.authorize(req,res,next), (req: Request, res: Response) => {
       res.redirect(/* 200, */ '/dash')
     })
-    app.post('/auth', async function(req,res) {
+    app.post('/auth', async function(req: Request, res: Response) {
         let username = req.body.username
         let password = req.body.password
         if(username==user && password==pass) {
+          //@ts-ignore
             req.session.user = username
+            //@ts-ignore
             req.session.pass = password
             res.redirect(/* 200, */ '/dash')
             }
@@ -142,7 +164,7 @@ res.send(await _this.renderFile(path.join(Theme.getPath(_this.theme), "index.ejs
     
     
     
-    app.get('/dash', islogin, async(req,res) => {
+    app.get('/dash', islogin, async(req: Request, res: Response) => {
     
     let user = await bot.users.fetch('694184230271451166')
      let author = user.username + "#" + user.discriminator
@@ -152,11 +174,11 @@ res.send(await _this.renderFile(path.join(Theme.getPath(_this.theme), "index.ejs
         })
         
     
-    app.get('/command', islogin, async(req,res) => {
+    app.get('/command', islogin, async(req: Request, res: Response) => {
         let text = ''
         return res.redirect('/status?code=501&r=Not%20Fixed')
  try{       
-async function *walkSync(dir) {
+async function *walkSync(dir:string):any {
   console.log(dir)
   const files = fs.readdirSync(dir.toString(), { withFileTypes: true });
   for (const file of files) {
@@ -164,9 +186,10 @@ async function *walkSync(dir) {
     if (file.isDirectory()) {
       yield* walkSync(path.join(dir, file.name.toString()));
     } else {
-      yield path.join(dir, file.name.toString());
+     yield path.join(dir, file.name.toString());
     }
   }
+  return [];
 }
 let ff = []
 for (const filePath of walkSync(command)) {
@@ -185,11 +208,11 @@ ${rr}</button></a></li>`*/
         catch(e) {
      text = "path is invalid or error occurred"
             }
-        res.send(await _this.renderFile(path.join(__dirname, 'views', 'command.ejs', { text, req })))
+        res.send(await _this.renderFile(path.join(__dirname, 'views', 'command.ejs'), { text, req }))
         })
     
     
-    app.get('/guild',islogin, async function (req,res) {
+    app.get('/guild',islogin, async function (req: Request, res: Response) {
         let guild = ''
         let server = bot.guilds.cache.toJSON()
  for(let i = 0;i<server.length;i++){
@@ -206,10 +229,12 @@ ${server[i].name}</button></a></li>`*/
         })
     
    
-    app.get('/guild/info', islogin, async function (req,res) {
+    app.get('/guild/info', islogin, async function (req: Request, res: Response) {
         let info = ''
         try {
-let guild = bot.guilds.cache.get(req.query.id)
+          if(!req.query.id) return res.redirect('/status?code=400&message=No%20ID%20Provided')
+let guild = bot.guilds.cache.get(req.query.id as string)
+if(!guild) return res.redirect('/status?code=404&message=Guild%20Not%20Found');
 info = `Id: ${guild.id}<br>Name: ${guild.name}<br>Owner Id: ${guild.ownerId}<br>Members count: ${guild.memberCount}<br>Features: ${guild.features.join(', ')}`
             }
         catch (e) {
@@ -220,16 +245,17 @@ info = `Id: ${guild.id}<br>Name: ${guild.name}<br>Owner Id: ${guild.ownerId}<br>
         })
     
     
-    app.get('/command/update',islogin, async function(req,res) {
-        bot.loader?.update()
+    app.get('/command/update',islogin, async function(req: Request, res: Response) {
+      //@ts-ignore 
+      bot.loader?.update()
         res.redirect(/* 200, */ '/command')
         })
     
-    app.get('/command/create', islogin, async function(req,res) {
+    app.get('/command/create', islogin, async function(req: Request, res: Response) {
         res.send(await _this.renderFile(path.join(__dirname, "views", "command_create.ejs")))
         }) 
     
-    app.post('/command/create', islogin, async function(req,res) {
+    app.post('/command/create', islogin, async function(req: Request, res: Response) {
         try{
         let nowname = command + '/' + req.body.name
         nowname = nowname.replace(/\//g, path.sep)
@@ -246,9 +272,10 @@ info = `Id: ${guild.id}<br>Name: ${guild.name}<br>Owner Id: ${guild.ownerId}<br>
     
     
     
-   app.get('/command/delete', islogin, async function(req,res) {
+   app.get('/command/delete', islogin, async function(req: Request, res: Response) {
        try {
-           let pathh = req.query.path
+           let pathh:string | null = req.query.path as string
+           if(!pathh) return res.redirect(`/status?code=400&message=No%20Path%20Provided`);
            pathh = pathh.replace(/%2F/g, path.sep)
            fs.unlinkSync(path.join(process.cwd(), pathh))
            res.redirect(/* 200, */ '/command')
@@ -259,9 +286,9 @@ info = `Id: ${guild.id}<br>Name: ${guild.name}<br>Owner Id: ${guild.ownerId}<br>
 
        })
     
-    app.get('/guild/leave', islogin, async function (req,res) {
+    app.get('/guild/leave', islogin, async function (req: Request, res: Response) {
         try { 
-            bot.guilds.cache.get(req.query.id).leave()
+            bot.guilds.cache.get(req.query.id as string)?.leave()
             res.redirect(/* 200, */ '/guild')
             }
         catch (e) {
@@ -272,18 +299,19 @@ info = `Id: ${guild.id}<br>Name: ${guild.name}<br>Owner Id: ${guild.ownerId}<br>
     
     
     
- app.get('/status' ,async function (req,res,next) {
-        res.status(req.query.code || 200)
+ app.get('/status' ,async function (req: Request, res: Response, next: Function) {
+  const statusCode:number = req.query.code? parseInt( req.query.code as string) :  200
+  res.status(statusCode)
        res.send(await _this.renderFile(path.join(__dirname, "views", "status.ejs")))
         })
     
-    app.get('/shell', islogin, async(req, res) => {
+    app.get('/shell', islogin, async(req: Request, res: Response) => {
       res.send(await _this.renderFile(path.join(__dirname, "views", "shell.ejs")))
 
     })
     
-    app.post('/shell', islogin, async(req, res) => {
-      let result = '';
+    app.post('/shell', islogin, async(req: Request, res: Response) => {
+      let result:any = '';
         try {
             result = await exec.execSync(`${req.body.execute}`).toString().replace(/\n/g, '<br>')
             }
@@ -296,12 +324,12 @@ info = `Id: ${guild.id}<br>Name: ${guild.name}<br>Owner Id: ${guild.ownerId}<br>
     // app.get('/login-c', async (req,res) => {
     //   res.send(await _this.renderFile(path.join(__dirname, 'views', "login_c.ejs")))
     // })
-    app.get('/djseval', islogin, async(req, res) => {
+    app.get('/djseval', islogin, async(req: Request, res: Response) => {
       res.send(await _this.renderFile(path.join(__dirname, "views", "djseval.ejs")))
 
     })
     
-    app.post('/djseval', islogin, async(req, res) => {
+    app.post('/djseval', islogin, async(req: Request, res: Response) => {
       let result;
 let text  = ("\n"+req.body.execute+"\n")
         try {
@@ -318,16 +346,16 @@ let text  = ("\n"+req.body.execute+"\n")
     })
     
     
-     app.get('/aoieval', islogin, async(req, res) => {
+     app.get('/aoieval', islogin, async(req: Request, res: Response) => {
       res.send(await _this.renderFile(path.join(__dirname, "views", "aoi_eval.ejs")))
 
     })
     
-    app.post('/aoieval', islogin, async(req, res) => {
+    app.post('/aoieval', islogin, async(req: Request, res: Response) => {
       let result;
         try {
             const client = bot
-      
+      //@ts-ignore
     result = await client.functionManager.interpreter(
                     client,
                     {},
@@ -350,7 +378,7 @@ let text  = ("\n"+req.body.execute+"\n")
     })
     
     
-   app.get('/reboot', islogin, async(req,res) => {
+   app.get('/reboot', islogin, async(req: Request, res: Response) => {
      
       await res.send(_this.renderFile(path.join(__dirname, "views", "reboot.ejs")))
 
@@ -366,7 +394,7 @@ let text  = ("\n"+req.body.execute+"\n")
         process.exit();
        })
     
-    async function islogin(req,res,next) {
+    async function islogin(req: Request, res: Response, next:Function) {
       // if(process.argv.length > 2) return next() 
       if(req.user && !req.url.includes('/owner')){
             return next()
@@ -395,8 +423,9 @@ let text  = ("\n"+req.body.execute+"\n")
     console.log("dashboard ready in port: "+port)
     }
 }
-
-module.exports = {
-  Dash,
-  Themes: Theme
-}
+export const Dashboard = Dash;
+// export default {
+//   Dash,
+//   Themes: Theme
+// }
+export const Themes = Theme;
